requirements.txt
pydantic>=2.0.0
google-generativeai
groq
python-dotenv
requests
beautifulsoup4
loguru
aiohttp
playwright>=1.40.0
playwright-stealth>=1.0.6
Pillow>=10.0.0
